requirements.txt
feedparser==6.0.11
python-dotenv==1.1.0
pytz==2025.2
uvloop==0.21.0
volcengine_python_sdk==1.1.2
lxml==5.3.1
httpx==0.28.1
pydantic==2.11.1